chore: format code & update files
