+++ tests/lesson-5/post-ver-2.spec.ts
@@ -1,132 +1,155 @@
 import { test, expect } from "@playwright/test";
 
-let xpathUserName = '//input[@id="user_login"]';
-let xpathPassword = '//input[@id="user_pass"]';
-let xpathBtnSubmit = '//input[@id="wp-submit"]';
-let xpathMenuPosts = "//div[text()='Posts']";
-let xpathMenuTags = "//a[text()='Tags']";
+let xpathUserName = "//input[@id='user_login']";
+let xpathPassword = "//input[@id='user_pass']";
+let xpathBtnLogin = "//input[@id='wp-submit']";
+let xpathMenuPosts = "//div[contains(text(),'Posts')]";
+let xpathMenuTags = "//a[contains(text(),'Tags')]";
+let xpathHeadingTags = "//h1[text()='Tags']";
 let xpathInputName = "//input[@id='tag-name']";
-let xpathBtnAddNewTag = "//input[@id='submit']";
 let xpathInputSlug = "//input[@id='tag-slug']";
+let xpathBtnAddNew = "//input[@id='submit']";
+let xpathMsgRequiredTagName = "//p[text()='A name is required for this term.']";
+let xpathMsgExisTagName = "//p[text()='A term with the name provided already exists in this taxonomy.']";
 let xpathMsgTagAdded = "//p[text()='Tag added.']";
-let xpathMsgRequired = "//p[text()='A name is required for this term.']";
-let xpathMsgNameExist = "//p[text()='A term with the name provided already exists in this taxonomy.']";
 let xpathMenuCategories = "//a[text()='Categories']";
-let xpathMesCategoryAdded = "//p[text()='Category added.']";
-let xpathBtnAddNewCategory = "//input[@value='Add New Category']";
+let xpathMsgCategoryAdded = "//p[text()='Category added.']";
 let xpathSelectParent = "//select[@id='parent']";
 
-const username = "1107-thu";
-const password = "HV%MhRjgyljzRDnxZCB(^Wl&";
+let username = "k11-trang";
+let password = "TCKoQJ4S3hKFyEamNgM0OwMK";
 
-function getXpathTagNameInTable(name: string) {
+function getXpathSlugInTable(slug: string) {
+    const convertToSlug = slug
+        .replace(/Đ/g, "D")
+        .replace(/đ/g, "d")
+        .normalize("NFD")
+        .replace(/[\u0300-\u036f]/g, "")
+        .toLowerCase()
+        .replace(/[^a-z0-9\s-]/g, "")
+        .trim()
+        .replace(/\s+/g, "-");
+    return `//td[text()='${convertToSlug}']`
+}
+
+function getXpathNameInTable(name: string) {
     return `//tbody//a[contains(text(),'${name}')]`;
 }
 
-function getXpathSlugNameInTable(slug: string) {
-    return `//td[text()='${slug}']`
-}
-
 function getXpathBtnDelete(name: string) {
-    return `//a[@aria-label='Delete “${name}”']`;
+    return `//a[@aria-label='Delete “${name}”']`
 }
 
-test.describe("POST - Post Tag failed", async () => {
+test.describe("POST - Post tag failed", async () => {
     test.beforeEach(async ({ page }) => {
-        await test.step("Go to menu Tags", async () => {
+        await test.step("Go to Tag page", async () => {
             await page.goto("https://pw-practice-dev.playwrightvn.com/wp-admin");
             await page.locator(xpathUserName).fill(username);
             await page.locator(xpathPassword).fill(password);
-            await page.locator(xpathBtnSubmit).click();
-            await page.locator(xpathMenuPosts).click();
-            await page.locator(xpathMenuTags).click();
+            await page.click(xpathBtnLogin);
+            await page.waitForLoadState("load");
+            await page.hover(xpathMenuPosts);
+            await page.click(xpathMenuTags);
+            await expect(page.locator(xpathHeadingTags)).toBeVisible();
         })
-    })
+    });
 
-    test("@POST_TAG_001 - Tag - add tag failed", async ({ page }) => {
-        await test.step("Click button 'Add New Tag'", async () => {
-            await page.locator(xpathBtnAddNewTag).click();
+    test("@POST_TAG_001: Tag - add tag failed", async ({ page }) => {
+        let existsName = "lesson tag";
 
-            await expect(page.locator(xpathMsgRequired)).toBeVisible();
-        })
+        await test.step("Click button [Add New Tag]", async () => {
+            await page.click(xpathBtnAddNew);
 
-        await test.step('Điền thông tin tag: name = "lesson tag", click button "Add New Tag"', async () => {
-            await page.locator(xpathInputName).fill("lesson tag");
-            await page.locator(xpathBtnAddNewTag).click();
-
-            await expect(page.locator(xpathMsgNameExist)).toBeVisible();
-        })
-    })
-})
-
-test.describe("POST - Post Tag success", async () => {
-    const name1 = "tag mentor";
-    const slug1 = "";
-    const name2 = "tag mentor 02";
-    const slug2 = "tag-mentor-02";
-
-    test.beforeEach(async ({ page }) => {
-        await test.step("Go to menu Tags", async () => {
-            await page.goto("https://pw-practice-dev.playwrightvn.com/wp-admin");
-            await page.locator(xpathUserName).fill(username);
-            await page.locator(xpathPassword).fill(password);
-            await page.locator(xpathBtnSubmit).click();
-            await page.locator(xpathMenuPosts).click();
-            await page.locator(xpathMenuTags).click();
-        })
-
-        page.on('dialog', async dialog => {
-            await dialog.accept();
-        })
-    })
-
-    test.afterEach(async ({ page }) => {
-        await page.locator(getXpathTagNameInTable(name1)).hover();
-        await page.locator(getXpathBtnDelete(name1)).click();
-
-        await expect(page.locator(getXpathTagNameInTable(name1))).not.toBeVisible();
-        await expect(page.locator(getXpathSlugNameInTable(slug1))).not.toBeVisible();
-
-        await page.locator(getXpathTagNameInTable(name2)).hover();
-        await page.locator(getXpathBtnDelete(name2)).click();
-
-        await expect(page.locator(getXpathTagNameInTable(name2))).not.toBeVisible();
-        await expect(page.locator(getXpathSlugNameInTable(slug2))).not.toBeVisible();
-    })
-
-    test("@POST_TAG_002 - Tag - add tag success", async ({ page }) => {
-        await test.step('Điền thông tin tag: name = "tag {name}" (name là tên bạn) Click button "Add New Tag"', async () => {
-            await page.locator(xpathInputName).fill(name1);
-            await page.locator(xpathBtnAddNewTag).click();
-
-            await expect(page.locator(xpathMsgTagAdded)).toBeVisible();
-            await expect(page.locator(getXpathTagNameInTable(name1))).toBeVisible();
-            // await expect(page.locator(getXpathSlugNameInTable(slug))).toBeVisible();
+            await expect(page.locator(xpathMsgRequiredTagName)).toBeVisible();
         });
 
-        await test.step('Điền thông tin tag: name = "tag {name} 02", slug = "tag-${name}-02" Click button "Add New Tag"', async () => {
-            await page.locator(xpathInputName).fill(name2);
-            await page.locator(xpathInputSlug).fill(slug2);
+        await test.step("Submit the already exists name", async () => {
+            await page.locator(xpathInputName).fill(existsName);
+            await page.click(xpathBtnAddNew);
 
-            await page.locator(xpathBtnAddNewTag).click();
+            await expect(page.locator(xpathMsgExisTagName)).toBeVisible();
+        })
+    });
+});
+
+test.describe("POST - Add Tag success", async () => {
+    let name1 = "tag Oanh Pham 01";
+    let name2 = "tag Oanh Pham 02";
+    let slug2 = "oanh-pham-02";
+    let name3 = "tag Oanh Pham 03";
+    let slug3 = "Đây là tag đặc biệt @100 $200";
+
+    test.beforeEach(async ({ page }) => {
+        await test.step("Go to Tag page", async () => {
+            await page.goto("https://pw-practice-dev.playwrightvn.com/wp-admin");
+            await page.locator(xpathUserName).fill(username);
+            await page.locator(xpathPassword).fill(password);
+            await page.click(xpathBtnLogin);
+            await page.waitForLoadState("load");
+            await page.hover(xpathMenuPosts);
+            await page.click(xpathMenuTags);
+            await expect(page.locator(xpathHeadingTags)).toBeVisible();
+        })
+    });
+
+    test.afterEach(async ({ page }) => {
+        page.on("dialog", async dialog => dialog.accept());
+
+        await page.locator(getXpathNameInTable(name1)).hover();
+        await page.locator(getXpathBtnDelete(name1)).click();
+        await expect(page.locator(getXpathNameInTable(name1))).toBeHidden();
+
+        await page.locator(getXpathNameInTable(name2)).hover();
+        await page.locator(getXpathBtnDelete(name2)).click();
+        await expect(page.locator(getXpathNameInTable(name2))).toBeHidden();
+
+        await page.locator(getXpathNameInTable(name3)).hover();
+        await page.locator(getXpathBtnDelete(name3)).click();
+        await expect(page.locator(getXpathNameInTable(name3))).toBeHidden();
+    });
+
+    test("@POST_TAG_002 & @POST_TAG_003: Tag - add tag success", async ({ page }) => {
+
+        await test.step("Submit valid name", async () => {
+            await page.locator(xpathInputName).fill(name1);
+            await page.click(xpathBtnAddNew);
 
             await expect(page.locator(xpathMsgTagAdded)).toBeVisible();
-            await expect(page.locator(getXpathTagNameInTable(name2))).toBeVisible();
-            await expect(page.locator(getXpathSlugNameInTable(slug2))).toBeVisible();
-        })
-    })
-})
+            await expect(page.locator(getXpathNameInTable(name1))).toBeVisible();
+        });
 
-test.describe("POST - Post Category", async () => {
+        await test.step("Submit valid name & slug", async () => {
+            await page.locator(xpathInputName).fill(name2);
+            await page.locator(xpathInputSlug).fill(slug2);
+            await page.click(xpathBtnAddNew);
+
+            await expect(page.locator(xpathMsgTagAdded)).toBeVisible();
+            await expect(page.locator(getXpathNameInTable(name2))).toBeVisible();
+            await expect(page.locator(getXpathSlugInTable(slug2))).toBeVisible();
+        });
+
+        await test.step("Submit special tag", async () => {
+            await page.locator(xpathInputName).fill(name3);
+            await page.locator(xpathInputSlug).fill(slug3);
+            await page.click(xpathBtnAddNew);
+
+            await expect(page.locator(xpathMsgTagAdded)).toBeVisible();
+            await expect(page.locator(getXpathNameInTable(name3))).toBeVisible();
+            await expect(page.locator(getXpathSlugInTable(slug3))).toBeVisible();
+        });
+    });
+});
+
+test.describe("POST - Category", async () => {
     let arrCategories = [
         {
-            name: "category mentor 03",
+            name: "category Oanh Pham 03",
             slug: "Đây là category đặc biệt @100 $200",
             expect: "day-la-category-dac-biet-100-200",
             parent: ""
         },
         {
-            name: "category mentor 04",
+            name: "category Oanh Pham 04",
             slug: "",
             expect: "",
             parent: "k11 class"
@@ -134,42 +157,35 @@ test.describe("POST - Post Category", async () => {
     ]
 
     test.beforeEach(async ({ page }) => {
-        const username = "k8-ruby";
-        const password = "zi%$zl1IlK2zGhcQ2*A)2wKC";
-
-        await test.step("Go to menu Posts", async () => {
+        await test.step("Go to Category page", async () => {
             await page.goto("https://pw-practice-dev.playwrightvn.com/wp-admin");
             await page.locator(xpathUserName).fill(username);
             await page.locator(xpathPassword).fill(password);
-            await page.click(xpathBtnSubmit);
-            await page.click(xpathMenuPosts);
-        })
-
-        page.on('dialog', async dialog => {
-            await dialog.accept();
-        })
-    })
-
-    test.afterEach(async ({ page }) => {
-        for (let i = 0; i < arrCategories.length; i++) {
-            await test.step("Clear data test", async () => {
-                await page.locator(getXpathTagNameInTable(arrCategories[i].name)).hover();
-                await page.locator(getXpathBtnDelete(arrCategories[i].name)).click();
-
-                // verify delete success
-                await page.reload();
-                await expect(page.locator(getXpathTagNameInTable(arrCategories[i].name))).not.toBeVisible();
-            })
-        }
-    })
-
-    test("@POST_CATEGORY_001 - Category - create category success", async ({ page }) => {
-        await test.step("Go to Category page", async () => {
+            await page.click(xpathBtnLogin);
+            await page.waitForLoadState("load");
+            await page.hover(xpathMenuPosts);
             await page.click(xpathMenuCategories);
         })
+    });
+
+    test.afterEach(async ({ page }) => {
+        page.on("dialog", async dialog => dialog.accept());
 
         for (let i = 0; i < arrCategories.length; i++) {
-            await test.step("Fill tag name and add new tag", async () => {
+            await test.step("Remove category", async () => {
+                await page.locator(getXpathNameInTable(arrCategories[i].name)).hover();
+                await page.locator(getXpathBtnDelete(arrCategories[i].name)).click();
+
+                // Verify delete success 
+                await page.reload();
+                await expect(page.locator(getXpathNameInTable(arrCategories[i].name))).not.toBeVisible();
+            })
+        }
+    });
+
+    test("@POST_CATEGORY_001: Category - create category success", async ({ page }) => {
+        for (let i = 0; i < arrCategories.length; i++) {
+            await test.step("Add new category", async () => {
                 await page.fill(xpathInputName, arrCategories[i].name);
                 if (arrCategories[i].slug != "") {
                     await page.fill(xpathInputSlug, arrCategories[i].slug);
@@ -179,17 +195,17 @@ test.describe("POST - Post Category", async () => {
                     await page.selectOption(xpathSelectParent, arrCategories[i].parent);
                 }
 
-                await page.click(xpathBtnAddNewCategory);
+                await page.click(xpathBtnAddNew);
 
-                // verify msg success
-                await expect(page.locator(xpathMesCategoryAdded)).toBeVisible();
+                // Verify msg success
+                await expect(page.locator(xpathMsgCategoryAdded)).toBeVisible();
 
-                // verify tag name in list tag
-                await expect(page.locator(getXpathTagNameInTable(arrCategories[i].name))).toBeVisible({ timeout: 5000 });
+                // Verify tag name in list tag
+                await expect(page.locator(getXpathNameInTable(arrCategories[i].name))).toBeVisible({ timeout: 5000 });
 
-                // verify slug name in list tag
+                // Verify slug name in list tag
                 if (arrCategories[i].slug != "") {
-                    await expect(page.locator(getXpathSlugNameInTable(arrCategories[i].expect))).toBeVisible();
+                    await expect(page.locator(getXpathSlugInTable(arrCategories[i].expect))).toBeVisible();
                 }
             })
         }
